requirements.txt
Flask==2.2
Flask_SocketIO==5.1.1
MarkupSafe==2.1.2 # Required by Azure
flask_jwt_extended==4.3.1
flask-restful==0.3.8
python-engineio==4.3.1
python-socketio==5.5.2
msal==1.17.0
Flask-Session==0.4.0
werkzeug==2.2.3
imutils==0.5.4
matplotlib==3.5.1
protobuf==3.20.0 # Deal with Python 4.21
moviepy==1.0.3
azure-storage-blob==12.9.0
tqdm==4.62.3
azure-search-documents==11.4.0
identity==0.2.0
python-dotenv